chore(connect-common): config to connect-common
With the goal of making connect-web independent of @trezor/connect this
is one step forward moving connect config to connect-common.
