+++ packages/connect-common/package.json
@@ -52,6 +52,7 @@
         "tslib": "^2.6.2"
     },
     "devDependencies": {
+        "@trezor/device-utils": "workspace:^",
         "@trezor/protobuf": "workspace:^",
         "@trezor/protocol": "workspace:^",
         "@types/chrome": "^0.0.299",
