+++ packages/connect-common/src/constants/index.ts
@@ -1,2 +1,3 @@
 export * as ERRORS from './errors';
 export * from './webextension';
+export * from './transport';
