+++ packages/connect-common/src/data/config.ts
@@ -1,31 +1,5 @@
-// origin: https://github.com/trezor/connect/blob/develop/src/data/config.json
-
-import { DeviceModelInternal } from '@trezor/device-utils';
-import { TREZOR_USB_DESCRIPTORS } from '@trezor/transport/src/constants';
-
-type Config = {
-    webusb: typeof TREZOR_USB_DESCRIPTORS;
-    whitelist: Array<{ origin: string; priority: number }>;
-    management: Array<{ origin: string }>;
-    knownHosts: Array<{ origin: string; label: string }>;
-    onionDomains: Record<string, string>;
-    supportedBrowsers: Record<
-        string,
-        {
-            version: number;
-            download: string;
-            update: string;
-        }
-    >;
-    supportedFirmware: Array<{
-        coin?: string[]; // Todo: better type?
-        capabilities?: string[]; // Todo: better type?
-        methods?: string[]; // Todo: better type?
-        min: Partial<Record<DeviceModelInternal, string>>;
-        max?: undefined; // NOTE: max field is not used anywhere at the moment, it is here for type compatibility
-        comment?: string[];
-    }>;
-};
+import { TREZOR_USB_DESCRIPTORS } from '../constants';
+import { Config } from '../types/config';
 
 export const config: Config = {
     webusb: TREZOR_USB_DESCRIPTORS,
