+++ packages/connect-common/tsconfig.libESM.json
@@ -14,6 +14,9 @@
             "path": "../utils"
         },
         {
+            "path": "../device-utils"
+        },
+        {
             "path": "../protobuf"
         },
         {
