+++ packages/connect-explorer/src/pages/index.mdx
@@ -476,7 +476,7 @@ export const ExampleHeading = styled.h3`
 
                 Trezor Connect Manifest requires that you, as a Trezor Connect integrator, share your email and application URL with us. This provides us with the ability to reach you in case of any required maintenance. This subscription is mandatory.
 
-                To ensure your extension is displayed with its name rather than its ID, you need to open a Pull Request to include it in the `knownHosts` section of [our configuration file](https://github.com/trezor/trezor-suite/blob/develop/packages/connect/src/data/config.ts#L17).
+                To ensure your extension is displayed with its name rather than its ID, you need to open a Pull Request to include it in the `knownHosts` section of [our configuration file](https://github.com/trezor/trezor-suite/blob/develop/packages/connect-common/src/data/config.ts#L17).
 
                 If you need more customization, refer to [init method documentation](./methods/other/init)
