+++ packages/connect-web/src/webextension/extensionPermissions.ts
@@ -1,7 +1,7 @@
 // origin: https://github.com/trezor/connect/blob/develop/src/js/webusb/extensionPermissions.js
 
-import { config } from '@trezor/connect/src/data/config';
 import { WEBEXTENSION } from '@trezor/connect-common/src/constants/webextension';
+import { config } from '@trezor/connect-common/src/data/config';
 
 // This file is hosted on https://connect.trezor.io/*/extension-permissions.html
 // It's included WITHIN webextension application in trezor-usb-permissions.html to allow pairing webusb and trezor.io domain properly.
