+++ packages/connect-web/src/webusb/index.ts
@@ -1,6 +1,6 @@
 // origin: https://github.com/trezor/connect/blob/develop/src/js/webusb/index.js
 
-import { config } from '@trezor/connect/src/data/config';
+import { config } from '@trezor/connect-common/src/data/config';
 
 // handle message received from connect.js
 const onload = () => {
