+++ packages/connect/src/api/common/paramsValidator.ts
@@ -1,9 +1,9 @@
 // origin: https://github.com/trezor/connect/blob/develop/src/js/core/methods/helpers/paramsValidator.js
 import { ERRORS } from '@trezor/connect-common/src/constants';
+import { config } from '@trezor/connect-common/src/data/config';
 import type { DeviceModelInternal } from '@trezor/device-utils';
 import { typedObjectKeys, versionUtils } from '@trezor/utils';
 
-import { config } from '../../data/config';
 import type { CoinInfo, FirmwareRange } from '../../types';
 import { fromHardened } from '../../utils/pathUtils';
