+++ packages/connect/src/utils/deviceFeaturesUtils.ts
@@ -1,8 +1,8 @@
+import { config } from '@trezor/connect-common/src/data/config';
 import { DeviceModelInternal, getFirmwareOrBootloaderVersionArray } from '@trezor/device-utils';
 import { MessagesSchema as PROTO } from '@trezor/protobuf';
 import { isArrayMember, versionUtils } from '@trezor/utils';
 
-import { config } from '../data/config';
 import { CoinInfo, Features, UnavailableCapabilities } from '../types';
 
 const DEFAULT_CAPABILITIES_T1: PROTO.Capability[] = [
