+++ packages/transport/src/index.ts
@@ -1,7 +1,7 @@
 export * as TRANSPORT_ERROR from './errors';
 
 export type { Descriptor, Session, MessageResponse } from './types';
-export { TREZOR_USB_DESCRIPTORS, TRANSPORT } from './constants';
+export { TRANSPORT } from './constants';
 
 export { AbstractTransport as Transport, isTransportInstance } from './transports/abstract';
 export { AbstractApiTransport } from './transports/abstractApi';
