+++ packages/transport/tsconfig.lib.json
@@ -7,6 +7,9 @@
     "include": ["./src"],
     "references": [
         {
+            "path": "../connect-common"
+        },
+        {
             "path": "../protobuf"
         },
         {
