+++ packages/transport/tsconfig.libESM.json
@@ -7,6 +7,9 @@
     },
     "references": [
         {
+            "path": "../connect-common"
+        },
+        {
             "path": "../protobuf"
         },
         {
